# Necessary changes on the BitVM side for the new execution engine