This PR includes the necessary changes on the BitVM side to use the candidate new engine discussed in https://github.com/BitVM/rust-bitcoin-scriptexec/pull/1.

Needs to be merged only after the new engine is merged to main.