@weikengchen 2024-04-05
Some benchmark:
fq12.inv 
original execution: 5.45s
new execution: 0.66s

scalar_mul (note: this implementation failed the test)
original execution: 205s 
new execution: 29.7s